feat: use UUIDv7 primary keys
UUID Version 7 (UUIDv7) is a time-ordered UUID which encodes a Unix
timestamp with millisecond precision in the most significant 48 bits. As
with all UUID formats, 6 bits are used to indicate the UUID version and
variant. The remaining 74 bits are randomly generated. As UUIDv7 is
time-ordered, values generated are practically sequential and therefore
eliminates the index locality problem.

The time-ordered nature of UUIDv7 results in much better DB performance
compared to random-prefixed UUIDv4s.